feature: Correccion y simplificacion del controlador, eliminacion de archivos de pruebas. udistrital/Cumplidos_y_certificaciones_documentacion#9
