+++ swagger/swagger.json
@@ -79,95 +79,6 @@
                 }
             }
         },
-        "/filtros_dependencia/": {
-            "post": {
-                "tags": [
-                    "filtros_dependencia"
-                ],
-                "description": "filter contracts and periods of validity by units and periods of validity",
-                "operationId": "FiltrosDependenciasController.FiltrarPorDependencias",
-                "parameters": [
-                    {
-                        "in": "body",
-                        "name": "body",
-                        "description": "body for units content",
-                        "required": true,
-                        "schema": {
-                            "$ref": "#/definitions/BodyParams"
-                        }
-                    }
-                ],
-                "responses": {
-                    "201": {
-                        "description": "{int} models.ContratoDependencia"
-                    },
-                    "403": {
-                        "description": "body is empty"
-                    }
-                }
-            }
-        },
-        "/filtros_pagos/": {
-            "get": {
-                "tags": [
-                    "filtros_pagos"
-                ],
-                "description": "get PagoMensual",
-                "operationId": "FiltrosPagosMensualesController.GetPagos",
-                "parameters": [
-                    {
-                        "in": "query",
-                        "name": "query",
-                        "description": "Filter. e.g. col1:v1,col2:v2 ...",
-                        "type": "string"
-                    },
-                    {
-                        "in": "query",
-                        "name": "fields",
-                        "description": "Fields returned. e.g. col1,col2 ...",
-                        "type": "string"
-                    },
-                    {
-                        "in": "query",
-                        "name": "sortby",
-                        "description": "Sorted-by fields. e.g. col1,col2 ...",
-                        "type": "string"
-                    },
-                    {
-                        "in": "query",
-                        "name": "order",
-                        "description": "Order corresponding to each sortby field, if single value, apply to all sortby fields. e.g. desc,asc ...",
-                        "type": "string"
-                    },
-                    {
-                        "in": "query",
-                        "name": "limit",
-                        "description": "Limit the size of result set. Must be an integer",
-                        "type": "string"
-                    },
-                    {
-                        "in": "query",
-                        "name": "offset",
-                        "description": "Start position of result set. Must be an integer",
-                        "type": "string"
-                    }
-                ],
-                "responses": {
-                    "200": {
-                        "description": "",
-                        "schema": {
-                            "type": "array",
-                            "items": {
-                                "$ref": "#/definitions/models.PagoMensual"
-                            }
-                        }
-                    },
-                    "400": {
-                        "description": "not found resource"
-                    }
-                }
-            }
-        },
         "/informacion_informe/preliquidacion/{pago_mensual_id}": {
             "get": {
                 "tags": [
@@ -766,36 +677,32 @@
         }
     },
     "definitions": {
-        "23578.0xc0004ef590.false": {
+        "23578.0xc000598b70.false": {
             "title": "false",
             "type": "object"
         },
-        "23858.0xc0004ef5c0.false": {
+        "23858.0xc000598ba0.false": {
             "title": "false",
             "type": "object"
         },
-        "23919.0xc0004ef5f0.false": {
+        "23919.0xc000598bd0.false": {
             "title": "false",
             "type": "object"
         },
-        "23987.0xc0004ef6e0.false": {
+        "23987.0xc000598cc0.false": {
             "title": "false",
             "type": "object"
         },
-        "24124.0xc0004ef710.false": {
+        "24124.0xc000598cf0.false": {
             "title": "false",
             "type": "object"
         },
-        "24275.0xc0004ef740.false": {
+        "24275.0xc000598d20.false": {
             "title": "false",
             "type": "object"
         },
-        "40471.0xc000558f90.false": {
+        "40471.0xc0005f0570.false": {
             "title": "false",
-            "type": "object"
-        },
-        "BodyParams": {
-            "title": "BodyParams",
             "type": "object"
         },
         "models.ActividadEspecifica": {
@@ -1270,13 +1177,13 @@
                     "type": "string"
                 },
                 "CDP": {
-                    "$ref": "#/definitions/23858.0xc0004ef5c0.false"
+                    "$ref": "#/definitions/23858.0xc000598ba0.false"
                 },
                 "Dependencia": {
                     "type": "string"
                 },
                 "EjecutadoDinero": {
-                    "$ref": "#/definitions/24275.0xc0004ef740.false"
+                    "$ref": "#/definitions/24275.0xc000598d20.false"
                 },
                 "FechaCPS": {
                     "type": "string"
@@ -1293,22 +1200,22 @@
                     "$ref": "#/definitions/models.FechasConNovedades"
                 },
                 "InformacionContratista": {
-                    "$ref": "#/definitions/24124.0xc0004ef710.false"
+                    "$ref": "#/definitions/24124.0xc000598cf0.false"
                 },
                 "Novedades": {
-                    "$ref": "#/definitions/23987.0xc0004ef6e0.false"
+                    "$ref": "#/definitions/23987.0xc000598cc0.false"
                 },
                 "Objeto": {
                     "type": "string"
                 },
                 "RP": {
-                    "$ref": "#/definitions/23919.0xc0004ef5f0.false"
+                    "$ref": "#/definitions/23919.0xc000598bd0.false"
                 },
                 "Sede": {
                     "type": "string"
                 },
                 "Supervisor": {
-                    "$ref": "#/definitions/23578.0xc0004ef590.false"
+                    "$ref": "#/definitions/23578.0xc000598b70.false"
                 },
                 "ValorContrato": {
                     "type": "integer",
@@ -1659,7 +1566,7 @@
                     "format": "datetime"
                 },
                 "Periodo": {
-                    "$ref": "#/definitions/40471.0xc000558f90.false"
+                    "$ref": "#/definitions/40471.0xc0005f0570.false"
                 }
             }
         }
